+++ cli/Tupfile.lua
@@ -22,7 +22,7 @@ tup.rule({"tuf.go.tmpl"},
 vpkg = "github.com/flynn/flynn/pkg/version"
 for i, os in ipairs({"darwin", "linux"}) do
   for j, arch in ipairs({"amd64", "386"}) do
-    tup.rule({"../installer/bindata.go", "../util/release/flynn-release", "tuf.go"},
+    tup.rule({"../installer/bindata.go", "tuf.go"},
              "^c go build %o^ GOOS="..os.." GOARCH="..arch.." ../util/_toolchain/go/bin/go build -o %o -ldflags=\"-X "..vpkg..".commit $GIT_COMMIT -X "..vpkg..".branch $GIT_BRANCH -X "..vpkg..".tag $GIT_TAG -X "..vpkg..".dirty $GIT_DIRTY\"",
              {string.format("bin/flynn-%s-%s", os, arch)})
   end
